Add script to publish the distribution
This will publish the distribution contents to the gh-pages branch. The
idea here is that no manual editing of the gh-pages branch is necessary.
The page build generator is used to build the website, README.md provides
the content for the generated page, including link buttons that match the
Architect theme. And by running this script, the distribution contents will
be added/updated.