Move Ol3 to Cesium conversion functions to a class
With this change, a user can easily override any step of the
conversion process by subclassing the base class.

Subclassing requires the user to compile the subclass and the library
together using the closure compiler. The reason is that the method calls
inside the base class are optimized by the compiler and would not match the
overrides of the subclass. For details see the '@export' documentation
from the closure compiler.

As a bonus, this change gets ride of 'olcs.core.glAliasedLineWidthRange'
and have core.js go on a diet.

See the CHANGES.md for a possible way to migrate existing code.