Cleanup unnecessary or broken examples
The boundless WMS server we used to use is broken, the CORS headers
are now missing. Use OSM tiles instead.

It was also used for the EPSG:4326 example. I searched for another WMS server
to switch to but could not find one. Since this example is not fundamental
(everything is handled automatically in OL3-Cesium) then I just remove it.

It was also used in the two synthetic layer examples. I switched to plain
OSM tiles for the example illustrating how to get good performances and
took the opportunity to remove the other extra example.

I kept the boundless server in the rastersync example for now.