Always rely on frameState to create attributions
Some TileSources (TileUTFGrid, TileJSON and BingMaps, to date) define
their `setAttributions` function with a `frameState` argument,
provided by PluggableMap's POSTRENDER event in OpenLayers. With
ol-cesium, OLImageProvider calls `setAttributions` on instantiation,
without any arguments -- which, of course, results in an exception.

In this commit, we've implemented getTileCredits in order to
dynamically generate credits based on what's in the viewport.

An example using Bing (with the same credentials in OpenLayer's
repository) has been added to show the change in action.