Fix conversion of icon anchor to pixelOffset
In Cesium the offset of an entity is specified in pixels from it's center by
default. Openlayers uses the top left position as origin.
While the origin is already converted since #653, this does not consider any
scaling of the image that happens in Cesium at a later stage through the scale
parameter. With a scale < 1 this can result in images that are off by multiple
times their own size.

Multiplying the pixel offset with the scale of the image resolves the issue.