Port to CesiumJS 1.121
The selection example is broken with CesiumJS >= 1.116:
the globe is all black or grey and does not react to navigation movements.
There is no error message.

After investigations, it appears to be caused by the last multipolygon of the Antartica country.
Since all other multipolygons are working flowlessly we just remove that one from the dataset.